add a default `ghost.db` file
This is because the running environment for Now is read-only,
and thus Ghost can't create this file if it's not there.

This makes Ghost run in read-only mode which is fine for this
little demo. We emphasize that you set up the `config.js`
file to use "mysql" or "postgres" as the backend for a more
permanent deployment (so that it can persist the data
throughout multiple deployments and version upgrades).